add labels

+++ src/components/card-layout/card-layout.component.ts
@@ -1,17 +1,13 @@
-import {
-  Component,
-  ElementRef,
-  Input,
-  Output,
-  ViewChild,
-  inject,
-} from '@angular/core';
+import { Component, ElementRef, Input, ViewChild, inject } from '@angular/core';
 import { Card } from '../../types/card';
 import { AddIconComponent } from '../Icons/add-icon/add-icon.component';
 import { CardsService } from '../../services/cards.service';
 import { DeleteIconComponent } from '../Icons/delete-icon/delete-icon.component';
 import { OptionsIconComponent } from '../Icons/options-icon/options-icon.component';
 import { ColorPaletteComponent } from '../color-palette/color-palette.component';
+import { LabelIconComponent } from '../Icons/label-icon/label-icon.component';
+import { TasksService } from '../../services/tasks.service';
+import { TasksComponent } from '../tasks/tasks.component';
 
 @Component({
   selector: 'app-card-layout',
@@ -19,8 +15,10 @@ import { ColorPaletteComponent } from '../color-palette/color-palette.component'
   imports: [
     AddIconComponent,
     DeleteIconComponent,
+    LabelIconComponent,
     OptionsIconComponent,
     ColorPaletteComponent,
+    TasksComponent,
   ],
   template: `
     <dialog #newTaskDialog class="dialog rounded bg-slate-400">
@@ -46,18 +44,29 @@ import { ColorPaletteComponent } from '../color-palette/color-palette.component'
       </form>
     </dialog>
 
-    <div class="card text-black p-4 rounded" [style.backgroundColor]="bgColor">
+    <div
+      class="card text-black p-4 rounded"
+      [style.backgroundColor]="
+        card.bgColor === 'DEFAULT' ? '#94a3b8' : card.bgColor
+      "
+    >
       <header
         class="card-header font-bold mb-5 flex items-center justify-between relative"
       >
-        <span>{{ card.title }}</span>
+        <input
+          #cardTitle
+          type="text"
+          class="bg-transparent"
+          [value]="card.title"
+          (change)="changeTitle()"
+        />
         <app-options-icon
           class="w-6 cursor-pointer"
           (click)="showOptions = !showOptions"
         ></app-options-icon>
         @if (showOptions) {
         <div
-          class="bg-white p-4 absolute top-6 -right-12 rounded flex flex-col gap-4"
+          class="bg-white p-4 absolute top-6 -right-12 rounded flex flex-col gap-4 z-10"
         >
           <button
             class="flex items-center gap-2 justify-center"
@@ -68,29 +77,16 @@ import { ColorPaletteComponent } from '../color-palette/color-palette.component'
           </button>
           <hr />
           <app-color-palette
-            (selectedColor)="(bgColor = $event) && (showOptions = false)"
+            (selectedColor)="setBgColor($event)"
           ></app-color-palette>
         </div>
         }
       </header>
-
-      <div class="card-tasks flex flex-col gap-3">
-        @for (task of card.tasks; track $index) {
-        <div
-          class="flex items-center justify-between card-task p-2 bg-slate-200 rounded shadow-md"
-        >
-          <span>{{ task }}</span>
-          <app-delete-icon
-            class="w-5 cursor-pointer"
-            (click)="deleteTask($index)"
-          />
-        </div>
-        }
-      </div>
+      <app-tasks [card]="card"></app-tasks>
 
       <footer
         class="card-footer mt-4 flex items-center gap-2 hover:cursor-pointer text-gray-800"
-        (click)="showModal()"
+        (click)="showTaskModal()"
       >
         <app-add-icon class="w-6 h-6" class="w-5" />
         <span>Add new task</span>
@@ -102,30 +98,54 @@ import { ColorPaletteComponent } from '../color-palette/color-palette.component'
 export class CardLayoutComponent {
   @Input({ required: true }) card!: Card;
   @ViewChild('newTaskDialog') dialog!: ElementRef<HTMLDialogElement>;
+  @ViewChild('cardTitle') input!: ElementRef<HTMLInputElement>;
 
   cardsService = inject(CardsService);
+  tasksService = inject(TasksService);
   showOptions: boolean = false;
-  bgColor: string = '#94a3b8';
 
-  showModal(): void {
+  showTaskModal(): void {
     this.dialog.nativeElement.showModal();
+  }
+
+  changeTitle(): void {
+    const { value } = this.input.nativeElement;
+    if (!value) return;
+
+    this.cardsService.renameCard(this.card.id, value);
   }
 
   addNewTask(): void {
     const input = this.dialog.nativeElement.querySelector('input');
     if (!input) return;
 
-    const task = input.value;
-    this.cardsService.createTask(this.card.id, task);
+    // find the highest id and increment by 1
+    const id =
+      this.card.tasks.reduce((acc, task) => {
+        const id = parseInt(task.id);
+        return id > acc ? id : acc;
+      }, 0) + 1;
+
+    const task = {
+      id: id.toString(),
+      title: input.value,
+      labels: [],
+    };
+
+    this.tasksService.createTask(this.card.id, task);
     input.value = '';
     this.dialog.nativeElement.close();
   }
 
-  deleteTask(index: number): void {
-    this.cardsService.deleteTask(this.card.id, index);
+  deleteCard(): void {
+    const response = confirm('Are you sure you want to delete this card?');
+    if (!response) return;
+    this.cardsService.deleteCard(this.card.id);
   }
 
-  deleteCard(): void {
-    this.cardsService.deleteCard(this.card.id);
+  setBgColor(color: string): void {
+    this.showOptions = false;
+    this.card.bgColor = color;
+    this.cardsService.saveCards();
   }
 }
